chore/API: speed up edit/test feedback loop for llmapi module
Previously, it took ~6 seconds for a single edit/test/debug feedback
loop in the `llmapi` module. After this change, it's now 1-2s.

The reason the feedback loop was slow was that we depended
on the `//cmd/frontend/internal/modelconfig` target, which transitively
brings in `graphqlbackend` and all the migration code, which adds huge
overhead to Go link times. It was relatively easy to untangle this
dependency so I went ahead and removed it to boost my local feedback
loop.