migration: specify column for soft delete triggers (#64446)
We believe the migration to add tenant_id is failing on dotcom is due to
this trigger which runs for any possible change to a row. Instead it
only needs to run if the deleted_at column changes. So this migration
just adjusts each trigger to only run if deleted_at is changed.

We don't need to change the trigger function, since the trigger
functions only read deleted at.

If this doesn't resolve our issue, we will need to look into a statement
level trigger for all our triggers.

Test Plan: CI and then will test safely on larger instances.