migrations: commit between table locks and use when condition for triggers
The improvement to the triggers I tried to ship yesterday failed on
dotcom. This is because it has the same issue as the tenant_id column
where we can't acquire the ShareRowExclusiveLock for multiple important
tables in a transaction without deadlocking. We use the same technique
where we commit and restart the transaction. This is safe to do since
even if it fails, the migration is idempotent and partial migration
still works.

Additionally after reading the documentation on triggers more deeply
[1], I noticed we can add in a WHEN clause to triggers. We already have
this in the functions, but this should be more efficient.

I didn't want to remove the condition from the functions just yet, since
that makes the migration more risky under partial failure.

Test Plan: CI. Also added BEGIN and COMMIT to the migration than ran
locally with

  psql -f migrations/frontend/1723557836_faster_soft_delete_trigger_for_permissions/up.sql

Then inspected the DB state for repo and had a good looking trigger come
back

  \d+ repo
  ...snip
  Triggers:
    trig_delete_user_repo_permissions_on_repo_soft_delete AFTER UPDATE
    OF deleted_at ON repo FOR EACH ROW WHEN (new.deleted_at IS NOT NULL
    AND old.deleted_at IS NULL) EXECUTE FUNCTION
    delete_user_repo_permissions_on_repo_soft_delete()

Additionally before landing I am going to manually run this migration on
S2 first, then on dotcom to see how it behaves when there is lots of
traffic and scale.