siteadmin: Make recoverUsers idempotent
This fails with some database constraint error right now, we should instead first validate that the set of users passed down is indeed soft-deleted which is a prerequisite for recovering users.

Test plan: Added tests for the resolver.

## Changelog

The recoverUsers endpoint failed with a spurious error when some of the given users were already active. It is now idempotent.