migrations: give new ids to rewritten migrations
The tenant_id and the trigger migration have had a few iterations. By
no-opping the older versions and re-introducing them this will ensure
anywhere we continuously deploy will get the latest version of them.

We have already applied the frontend migration to s2 and dotcom, so it
should be a no-op for them.

Test Plan: CI. Will only land early tomorrow morning CET to monitor
potential impact. Additionally will check the state of the code-intel
and codeinsights DBs on dotcom before landing.