database: limit concurrent event log inserts
We have had downtime on dotcom twice this week due to a fast migration
that tries to acquire a lock on the event_logs table. In both cases we
had a huge backlog of events to write to the DB which ended up
exhausting the connection pool since they also don't respect context
cancellation.

This is a speculative fix for this. The default of 5 was just made up, I
didn't want to go too low and I know this still gives lots of space for
other connections in our pool. I don't really want to test it on dotcom
after it lands, just in case it doesn't work :)

Test Plan: CI to just check it still works.