codeintel: run uploads ExpireFailedRecords janitor job every 30m
I was inspecting the dotcom frontend database slow query log, and this
query by far was using the most clock time over the period. We should
look into making this run faster, but for now we can just not run it as
often. It's a janitor job, so not critical to run as often.

Test Plan: change is so simple that I will just see if this disappears
from the slow query log.