fix flakiness in TestPeriodicGoroutine
This is my best guess as to why this test was flakey. I believe it was
racing with goroutine.Start running the handler and the call to clock.

Either way, I reckon this will still be flakey and the only real way to
fix it is to be fine with either 3 or 4 calls (or after this change 4 or
5 calls).