feat/enterpriseportal: define SubscriptionLicenseChecksService (#64396)
This PR defines an Enterprise Portal version of the "license check" API
that is currently bundled in Souregraph as a separate gRPC service that
will be implemented in Enterprise Portal (in a PR I will stack on top of
this one). The API schema is pretty much the same, except:

- removed the `error` field, since I'm pretty sure we can return a real
error instead in a connectRPC implementation for the same effect.
- made `license_key` a parameter, as we already propagate the license
key all over the place (pings, telemetry) - this is a fairly
low-frequency check so there shouldn't be any bandwidth concerns or the
like. For now, this parameter will support the existing format so we can
forward checks in dotcom to EP, but I've set a target removal release

Existing API:
https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/cmd/frontend/internal/dotcom/productsubscription/license_check_handler.go?L133

Part of https://linear.app/sourcegraph/issue/CORE-227

## Test plan

CI