Gather additional statistics at index time (#64426)
Adds a new `GetCodyContextAlternatives` GraphQL resolver that returns a
list of `ContextList`, each of which is a distinct list of results that
match the user query, along with a name that describes how that result
list was computed. The idea here is to be able to show alternative
context fetched with different methods/rankers, which can be displayed
by the client to help us iterate on context quality and ordering.

The existing context fetching mechanism is preserved and given the
`keyword($query)` name.

We add a new experimental context list, which modifies the user query to
do term expansion using a table of keywords. This keyword table is
computed in the `RepositoryReindex` GraphQL endpoint and stored in
Redis. We do not worry about clean up, because this feature is
experimental.

All of this is protected by the feature flag `enhanced-index`. Existing
production behavior should not be modified.

---------

Co-authored-by: Rishabh Mehrotra <erishabh@gmail.com>