telemetry: always best-effort write to v1 store
In another PR addressing event_log tables being an issue, it was
mentioned that we can just write to the legacy store without monitoring
if it works or not. This commit implements that. Most of the code change
is just prop drilling for the logger. We could simplify even further and
not even log, but it feels like not being able to see the error anywhere
would be hard to debug.

Test Plan: CI